Trimming/Slicing input
Rather than analyzing entire length of the sample based on arbitrary value we now sample the exact length of sample size and only run analysis on the first few seconds. Stereo audio is spliced to mono for accuracy.

This data is then processed using a Hann window to filter out excessive signal. We also chop unnecessarily high frequencies for harmonics to make the process easier.
We downsample the WAV for a smaller signal input and less noise. CLI ouput is currently for debugging (trying to configure optimal window sizing/frequency ranges)